added logger

+++ src/main/java/controller/DeleteCarController.java
@@ -3,8 +3,10 @@ package controller;
 import model.entity.Car;
 import model.entity.User;
 import model.service.CarService;
-import model.service.UsefulFunctions;
+import model.dao.Dao;
 import model.service.UserService;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -15,29 +17,31 @@ import java.sql.SQLException;
 import java.util.List;
 
 public class DeleteCarController extends HttpServlet {
+    final static Logger logger = LogManager.getLogger(DeleteCarController.class);
     @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        if(getServletContext().getAttribute(UsefulFunctions.UserID)==null)
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
+        if(getServletContext().getAttribute(Dao.USER_ID)==null) {
             resp.sendRedirect("index.jsp");
-        Integer id = (Integer) getServletContext().getAttribute(UsefulFunctions.UserID);
+        }
+        Integer id = (Integer) getServletContext().getAttribute(Dao.USER_ID);
         try {
             User user = UserService.getUserById(id);
             String carNumber = req.getParameter("carNumber");
             List<Car> cars = user.getCars();
             for(int i = 0; i<cars.size(); i++){
-                if(cars.get(i).getCarNumber().equals(carNumber)) {
-                    try {
-                        CarService.deleteCar(cars.get(i));
-                    } catch (SQLException | ClassNotFoundException e) {
-                        e.printStackTrace();
-                    }
+                if(cars.get(i).getCarNumber().equals(carNumber) && !cars.get(i).isParked()) {
+                    CarService.deleteCar(cars.get(i));
+                    logger.info("car with number "+carNumber+" deleted, user "+id);
                     break;
                 }
+                if(i==cars.size()-1){
+                    logger.info(" couldn't delete car with number "+carNumber+", user "+id);
+                }
             }
-            resp.sendRedirect("account");
         } catch (SQLException | ClassNotFoundException e) {
-            e.printStackTrace();
+            logger.error("DB exception, user "+id);
         }
+        resp.sendRedirect("account");
     }
 
     @Override
